retun type handling

+++ goamdsmi.go
@@ -51,27 +51,27 @@ package goamdsmi
 import "C"
 
 //GPU ROCM or AMDSMI calls
-func GO_gpu_init() (uint) {
-	return uint(C.goamdsmi_gpu_init())
+func GO_gpu_init() (int) {
+	return int(C.goamdsmi_gpu_init())
 }
 
-func GO_gpu_shutdown() (uint) {
-	return uint(C.goamdsmi_gpu_shutdown())
+func GO_gpu_shutdown() (int) {
+	return int(C.goamdsmi_gpu_shutdown())
 }
 
-func GO_gpu_num_monitor_devices() (uint) {
-	return uint(C.goamdsmi_gpu_num_monitor_devices())
+func GO_gpu_num_monitor_devices() (int) {
+	return int(C.goamdsmi_gpu_num_monitor_devices())
 }
 
 func GO_gpu_dev_name_get(i int) (*C.char) {
 	return C.goamdsmi_gpu_dev_name_get(C.uint(i))
 }
 
-func GO_gpu_dev_id_get(i int) (C.uint16_t) {
-	return C.uint16_t(C.goamdsmi_gpu_dev_id_get(C.uint(i)))
+func GO_gpu_dev_id_get(i int) (C.int) {
+	return C.int(C.goamdsmi_gpu_dev_id_get(C.uint(i)))
 }
 
-func GO_gpu_dev_pci_id_get(i int) (C.uint64_t) {
+func GO_gpu_dev_pci_id_get(i int) (C.int) {
 	return C.goamdsmi_gpu_dev_pci_id_get(C.uint(i))
 }
 
@@ -83,108 +83,108 @@ func GO_gpu_dev_vendor_name_get(i int) (*C.char) {
 	return C.goamdsmi_gpu_dev_vendor_name_get(C.uint(i))
 }
 
-func GO_gpu_dev_power_cap_get(i int) (C.uint64_t) {
+func GO_gpu_dev_power_cap_get(i int) (C.int) {
     return C.goamdsmi_gpu_dev_power_cap_get(C.uint(i))
 }
 
-func GO_gpu_dev_power_ave_get(i int) (C.uint64_t) {
+func GO_gpu_dev_power_ave_get(i int) (C.int) {
     return C.goamdsmi_gpu_dev_power_ave_get(C.uint(i))
 }
 
-func GO_gpu_dev_temp_metric_get(i int, sensor int, metric int) (C.uint64_t) {
+func GO_gpu_dev_temp_metric_get(i int, sensor int, metric int) (C.int) {
     return C.goamdsmi_gpu_dev_temp_metric_get(C.uint(i), C.uint(sensor), C.uint(metric))
 }
 
-func GO_gpu_dev_perf_level_get(i int) (C.uint32_t) {
+func GO_gpu_dev_perf_level_get(i int) (C.int) {
 	return C.goamdsmi_gpu_dev_perf_level_get(C.uint(i))
 }
 
-func GO_gpu_dev_overdrive_level_get(i int) (C.uint32_t) {
+func GO_gpu_dev_overdrive_level_get(i int) (C.int) {
 	return C.goamdsmi_gpu_dev_overdrive_level_get(C.uint(i))
 }
 
-func GO_gpu_dev_mem_overdrive_level_get(i int) (C.uint32_t) {
+func GO_gpu_dev_mem_overdrive_level_get(i int) (C.int) {
 	return C.goamdsmi_gpu_dev_mem_overdrive_level_get(C.uint(i))
 }
 
-func GO_gpu_dev_gpu_clk_freq_get_sclk(i int) (C.uint64_t) {
+func GO_gpu_dev_gpu_clk_freq_get_sclk(i int) (C.int) {
 	return C.goamdsmi_gpu_dev_gpu_clk_freq_get_sclk(C.uint(i))
 }
 
-func GO_gpu_dev_gpu_clk_freq_get_mclk(i int) (C.uint64_t) {
+func GO_gpu_dev_gpu_clk_freq_get_mclk(i int) (C.int) {
     return C.goamdsmi_gpu_dev_gpu_clk_freq_get_mclk(C.uint(i))
 }
 
-func GO_gpu_od_volt_freq_range_min_get_sclk(i int) (C.uint64_t) {
+func GO_gpu_od_volt_freq_range_min_get_sclk(i int) (C.int) {
 	return C.goamdsmi_gpu_od_volt_freq_range_min_get_sclk(C.uint(i))
 }
 
-func GO_gpu_od_volt_freq_range_min_get_mclk(i int) (C.uint64_t) {
+func GO_gpu_od_volt_freq_range_min_get_mclk(i int) (C.int) {
 	return C.goamdsmi_gpu_od_volt_freq_range_min_get_mclk(C.uint(i))
 }
 
-func GO_gpu_od_volt_freq_range_max_get_sclk(i int) (C.uint64_t) {
+func GO_gpu_od_volt_freq_range_max_get_sclk(i int) (C.int) {
 	return C.goamdsmi_gpu_od_volt_freq_range_max_get_sclk(C.uint(i))
 }
 
-func GO_gpu_od_volt_freq_range_max_get_mclk(i int) (C.uint64_t) {
+func GO_gpu_od_volt_freq_range_max_get_mclk(i int) (C.int) {
 	return C.goamdsmi_gpu_od_volt_freq_range_max_get_mclk(C.uint(i))
 }
 
-func GO_gpu_dev_gpu_busy_percent_get(i int) (C.uint64_t) {
+func GO_gpu_dev_gpu_busy_percent_get(i int) (C.int) {
 	return C.goamdsmi_gpu_dev_gpu_busy_percent_get(C.uint(i))
 }
 
-func GO_gpu_dev_gpu_memory_busy_percent_get(i int) (C.uint64_t) {
+func GO_gpu_dev_gpu_memory_busy_percent_get(i int) (C.int) {
 	return C.goamdsmi_gpu_dev_gpu_memory_busy_percent_get(C.uint(i))
 }
 
-func GO_gpu_dev_gpu_memory_usage_get (i int) (C.uint64_t) {
+func GO_gpu_dev_gpu_memory_usage_get (i int) (C.int) {
 	return C.goamdsmi_gpu_dev_gpu_memory_usage_get(C.uint(i))
 }
 
-func GO_gpu_dev_gpu_memory_total_get (i int) (C.uint64_t) {
+func GO_gpu_dev_gpu_memory_total_get (i int) (C.int) {
 	return C.goamdsmi_gpu_dev_gpu_memory_total_get(C.uint(i))
 }
 
 //CPU ESMI or AMDSMI calls
-func GO_cpu_init() (uint) {
-	return uint(C.goamdsmi_cpu_init())
+func GO_cpu_init() (int) {
+	return int(C.goamdsmi_cpu_init())
 }
 
-func GO_cpu_number_of_sockets_get() (uint) {
-	return uint(C.goamdsmi_cpu_number_of_sockets_get())
+func GO_cpu_number_of_sockets_get() (int) {
+	return int(C.goamdsmi_cpu_number_of_sockets_get())
 }
 
-func GO_cpu_number_of_threads_get() (uint) {
-	return uint(C.goamdsmi_cpu_number_of_threads_get())
+func GO_cpu_number_of_threads_get() (int) {
+	return int(C.goamdsmi_cpu_number_of_threads_get())
 }
 
-func GO_cpu_threads_per_core_get() (uint) {
-	return uint(C.goamdsmi_cpu_threads_per_core_get())
+func GO_cpu_threads_per_core_get() (int) {
+	return int(C.goamdsmi_cpu_threads_per_core_get())
 }
 
-func GO_cpu_core_energy_get(i int) (C.uint64_t) {
+func GO_cpu_core_energy_get(i int) (C.int) {
 	return C.goamdsmi_cpu_core_energy_get(C.uint(i))
 }
 
-func GO_cpu_core_boostlimit_get(i int) (C.uint32_t) {
+func GO_cpu_core_boostlimit_get(i int) (C.int) {
 	return C.goamdsmi_cpu_core_boostlimit_get(C.uint(i))
 }
 
-func GO_cpu_socket_energy_get(i int) (C.uint64_t) {
+func GO_cpu_socket_energy_get(i int) (C.int) {
 	return C.goamdsmi_cpu_socket_energy_get(C.uint(i))
 }
 
-func GO_cpu_socket_power_get(i int) (C.uint32_t) {
+func GO_cpu_socket_power_get(i int) (C.int) {
 	return C.goamdsmi_cpu_socket_power_get(C.uint(i))
 }
 
-func GO_cpu_socket_power_cap_get(i int) (C.uint32_t) {
+func GO_cpu_socket_power_cap_get(i int) (C.int) {
 	return C.goamdsmi_cpu_socket_power_cap_get(C.uint(i))
 }
 
-func GO_cpu_prochot_status_get(i int) (C.uint32_t) {
+func GO_cpu_prochot_status_get(i int) (C.int) {
 	return C.goamdsmi_cpu_prochot_status_get(C.uint(i))
 }
